+++ convex/grades.ts
@@ -1,9 +1,9 @@
 import { ConvexError, v } from 'convex/values'
 import { Doc } from './_generated/dataModel'
 import { mutation, query } from './_generated/server'
-import { requireAuthAndOwnership } from './authHelpers'
+import { requireAuth, requireAuthAndOwnership } from './authHelpers'
 import { gradeFields, gradeObject } from './schema'
-import { gradeSchema, validateWithSchema } from './validation'
+import { gradeSchema, validateGradeTotal, validateWithSchema } from './validation'
 
 /**
  * Add a grade to an assessment
@@ -26,10 +26,22 @@ export const addGrade = mutation({
     // Check if assessment exists and belongs to user
     const { userId, data: assessment } = await requireAuthAndOwnership(ctx, args.assessmentId)
 
+    // Get existing grades for this assessment to check total
     const existingGrades = await ctx.db
       .query('grades')
       .withIndex('by_assessment', (q) => q.eq('assessmentId', args.assessmentId))
       .collect()
+
+    const existingGradeValues = existingGrades.map((g) => g.grade)
+    const gradeValidation = validateGradeTotal(
+      validation.data.grade,
+      existingGradeValues,
+      (assessment as Doc<'assessments'>).weight,
+    )
+
+    if (!gradeValidation.isValid) {
+      throw new ConvexError(gradeValidation.error!)
+    }
 
     const gradeId = await ctx.db.insert('grades', {
       ...validation.data,
@@ -38,12 +50,32 @@ export const addGrade = mutation({
       assessmentId: args.assessmentId,
     })
 
-    // Update subject total grade
-    await ctx.db.patch(assessment.subjectId, {
-      totalGrade: existingGrades.reduce((sum, g) => sum + g.grade, 0) + validation.data.grade,
-    })
-
     return gradeId
+  },
+})
+
+/**
+ * Get all grades for a user
+ */
+export const getGradesByUser = query({
+  args: {
+    subjectId: v.optional(v.id('subjects')),
+  },
+  returns: v.array(gradeObject),
+  handler: async (ctx, args) => {
+    const userId = await requireAuth(ctx)
+
+    const hasSubjectFilter = !!args.subjectId
+
+    let results
+    results = ctx.db.query('grades').withIndex('by_user', (q) => q.eq('userId', userId))
+
+    // Apply subject filter
+    if (hasSubjectFilter) {
+      results = results.filter((q) => q.eq(q.field('subjectId'), args.subjectId))
+    }
+
+    return await results.collect()
   },
 })
 
@@ -52,7 +84,7 @@ export const addGrade = mutation({
  */
 export const getGradesByAssessment = query({
   args: {
-    assessmentId: gradeFields.assessmentId,
+    assessmentId: v.id('assessments'),
   },
   returns: v.array(gradeObject),
   handler: async (ctx, args) => {
@@ -99,25 +131,30 @@ export const updateGrade = mutation({
       }
     }
 
-    await ctx.db.patch(args.gradeId, validation.data)
-
-    // Update subject total grade if grade value was changed
+    // If grade is being updated, validate total doesn't exceed assessment weight
     if (validation.data.grade !== undefined) {
-      const allGrades = await ctx.db
+      // Get the assessment to check its weight
+      const assessment = await ctx.db.get(gradeRecord.assessmentId)
+      if (!assessment) {
+        throw new ConvexError('Assessment not found')
+      }
+
+      // Get existing grades for this assessment (excluding the current one)
+      const existingGrades = await ctx.db
         .query('grades')
         .withIndex('by_assessment', (q) => q.eq('assessmentId', gradeRecord.assessmentId))
+        .filter((q) => q.neq(q.field('_id'), args.gradeId))
         .collect()
 
-      // Find the updated grade in the collection and use its new value
-      const totalGrade = allGrades.reduce((sum, g) => {
-        if (g._id === args.gradeId) {
-          return sum + validation.data.grade!
-        }
-        return sum + g.grade
-      }, 0)
+      const existingGradeValues = existingGrades.map((g) => g.grade)
+      const gradeValidation = validateGradeTotal(validation.data.grade, existingGradeValues, assessment.weight)
 
-      await ctx.db.patch(gradeRecord.subjectId, { totalGrade })
+      if (!gradeValidation.isValid) {
+        throw new ConvexError(gradeValidation.error!)
+      }
     }
+
+    await ctx.db.patch(args.gradeId, validation.data)
   },
 })
 
@@ -130,18 +167,7 @@ export const deleteGrade = mutation({
   },
   returns: v.null(),
   handler: async (ctx, args) => {
-    const { data: gradeRecord } = await requireAuthAndOwnership(ctx, args.gradeId)
-
+    await requireAuthAndOwnership(ctx, args.gradeId)
     await ctx.db.delete(args.gradeId)
-
-    // Update subject total grade after deletion
-    const remainingGrades = await ctx.db
-      .query('grades')
-      .withIndex('by_assessment', (q) => q.eq('assessmentId', gradeRecord.assessmentId))
-      .collect()
-
-    const totalGrade = remainingGrades.reduce((sum, g) => sum + g.grade, 0)
-
-    await ctx.db.patch(gradeRecord.subjectId, { totalGrade })
   },
 })
