+++ package.json
@@ -66,7 +66,7 @@
     "usehooks-ts": "^3.1.1",
     "validator": "^13.15.15",
     "vaul": "^1.1.2",
-    "zod": "^4.0.15"
+    "zod": "^4.1.5"
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3",
